Made return values 8 byte unsigned integers

+++ src/Memory.cpp
@@ -39,41 +39,41 @@
  * memory use) measured in bytes, or zero if the value cannot be
  * determined on this OS.
  */
-size_t rlib::getPeakRSS( )
+u_int64_t rlib::getPeakRSS( )
 {
 #if defined(_WIN32)
     /* Windows -------------------------------------------------- */
     PROCESS_MEMORY_COUNTERS info;
     GetProcessMemoryInfo( GetCurrentProcess( ), &info, sizeof(info) );
-    return (size_t)info.PeakWorkingSetSize;
+    return (u_int64_t)info.PeakWorkingSetSize;
 
 #elif (defined(_AIX) || defined(__TOS__AIX__)) || (defined(__sun__) || defined(__sun) || defined(sun) && (defined(__SVR4) || defined(__svr4__)))
     /* AIX and Solaris ------------------------------------------ */
     struct psinfo psinfo;
     int fd = -1;
     if ( (fd = open( "/proc/self/psinfo", O_RDONLY )) == -1 )
-        return (size_t)0L;      /* Can't open? */
+        return (u_int64_t)0L;      /* Can't open? */
     if ( read( fd, &psinfo, sizeof(psinfo) ) != sizeof(psinfo) )
     {
         close( fd );
-        return (size_t)0L;      /* Can't read? */
+        return (u_int64_t)0L;      /* Can't read? */
     }
     close( fd );
-    return (size_t)(psinfo.pr_rssize * 1024L);
+    return (u_int64_t)(psinfo.pr_rssize * 1024L);
 
 #elif defined(__unix__) || defined(__unix) || defined(unix) || (defined(__APPLE__) && defined(__MACH__))
     /* BSD, Linux, and OSX -------------------------------------- */
     struct rusage rusage;
     getrusage( RUSAGE_SELF, &rusage );
 #if defined(__APPLE__) && defined(__MACH__)
-    return (size_t)rusage.ru_maxrss;
+    return (u_int64_t)rusage.ru_maxrss;
 #else
-    return (size_t)(rusage.ru_maxrss * 1024L);
+    return (u_int64_t)(rusage.ru_maxrss * 1024L);
 #endif
 
 #else
     /* Unknown OS ----------------------------------------------- */
-    return (size_t)0L;          /* Unsupported. */
+    return (u_int64_t)0L;          /* Unsupported. */
 #endif
 }
 
@@ -85,13 +85,13 @@ size_t rlib::getPeakRSS( )
  * Returns the current resident set size (physical memory use) measured
  * in bytes, or zero if the value cannot be determined on this OS.
  */
-size_t rlib::getCurrentRSS( )
+u_int64_t rlib::getCurrentRSS( )
 {
 #if defined(_WIN32)
     /* Windows -------------------------------------------------- */
     PROCESS_MEMORY_COUNTERS info;
     GetProcessMemoryInfo( GetCurrentProcess( ), &info, sizeof(info) );
-    return (size_t)info.WorkingSetSize;
+    return (u_int64_t)info.WorkingSetSize;
 
 #elif defined(__APPLE__) && defined(__MACH__)
     /* OSX ------------------------------------------------------ */
@@ -99,26 +99,26 @@ size_t rlib::getCurrentRSS( )
     mach_msg_type_number_t infoCount = MACH_TASK_BASIC_INFO_COUNT;
     if ( task_info( mach_task_self( ), MACH_TASK_BASIC_INFO,
         (task_info_t)&info, &infoCount ) != KERN_SUCCESS )
-        return (size_t)0L;      /* Can't access? */
-    return (size_t)info.resident_size;
+        return (u_int64_t)0L;      /* Can't access? */
+    return (u_int64_t)info.resident_size;
 
 #elif defined(__linux__) || defined(__linux) || defined(linux) || defined(__gnu_linux__)
     /* Linux ---------------------------------------------------- */
     long rss = 0L;
     FILE* fp = NULL;
     if ( (fp = fopen( "/proc/self/statm", "r" )) == NULL )
-        return (size_t)0L;      /* Can't open? */
+        return (u_int64_t)0L;      /* Can't open? */
     if ( fscanf( fp, "%*s%ld", &rss ) != 1 )
     {
         fclose( fp );
-        return (size_t)0L;      /* Can't read? */
+        return (u_int64_t)0L;      /* Can't read? */
     }
     fclose( fp );
-    return (size_t)rss * (size_t)sysconf( _SC_PAGESIZE);
+    return (u_int64_t)rss * (u_int64_t)sysconf( _SC_PAGESIZE);
 
 #else
     /* AIX, BSD, Solaris, and Unknown OS ------------------------ */
-    return (size_t)0L;          /* Unsupported. */
+    return (u_int64_t)0L;          /* Unsupported. */
 #endif
 }
 
@@ -143,7 +143,7 @@ size_t rlib::getCurrentRSS( )
 /**
  * Returns the size of physical memory (RAM) in bytes.
  */
-size_t rlib::getMemorySize( )
+u_int64_t rlib::getMemorySize( )
 {
 #if defined(_WIN32) && (defined(__CYGWIN__) || defined(__CYGWIN32__))
 	/* Cygwin under Windows. ------------------------------------ */
@@ -151,7 +151,7 @@ size_t rlib::getMemorySize( )
 	MEMORYSTATUS status;
 	status.dwLength = sizeof(status);
 	GlobalMemoryStatus( &status );
-	return (size_t)status.dwTotalPhys;
+	return (u_int64_t)status.dwTotalPhys;
 
 #elif defined(_WIN32)
 	/* Windows. ------------------------------------------------- */
@@ -159,7 +159,7 @@ size_t rlib::getMemorySize( )
 	MEMORYSTATUSEX status;
 	status.dwLength = sizeof(status);
 	GlobalMemoryStatusEx( &status );
-	return (size_t)status.ullTotalPhys;
+	return (u_int64_t)status.ullTotalPhys;
 
 #elif defined(__unix__) || defined(__unix) || defined(unix) || (defined(__APPLE__) && defined(__MACH__))
 	/* UNIX variants. ------------------------------------------- */
@@ -173,25 +173,25 @@ size_t rlib::getMemorySize( )
 #elif defined(HW_PHYSMEM64)
 	mib[1] = HW_PHYSMEM64;          /* NetBSD, OpenBSD. --------- */
 #endif
-	int64_t size = 0;               /* 64-bit */
-	size_t len = sizeof( size );
+	u_int64_t size = 0;              /* 64-bit */
+	u_int64_t len = sizeof( size );
 	if ( sysctl( mib, 2, &size, &len, NULL, 0 ) == 0 )
-		return (size_t)size;
-	return 0L;			/* Failed? */
+		return size;
+	return (u_int64_t)0L;			/* Failed? */
 
 #elif defined(_SC_AIX_REALMEM)
 	/* AIX. ----------------------------------------------------- */
-	return (size_t)sysconf( _SC_AIX_REALMEM ) * (size_t)1024L;
+	return (u_int64_t)sysconf( _SC_AIX_REALMEM ) * (u_int64_t)1024L;
 
 #elif defined(_SC_PHYS_PAGES) && defined(_SC_PAGESIZE)
 	/* FreeBSD, Linux, OpenBSD, and Solaris. -------------------- */
-	return (size_t)sysconf( _SC_PHYS_PAGES ) *
-		(size_t)sysconf( _SC_PAGESIZE );
+	return (u_int64_t)sysconf( _SC_PHYS_PAGES ) *
+		(u_int64_t)sysconf( _SC_PAGESIZE );
 
 #elif defined(_SC_PHYS_PAGES) && defined(_SC_PAGE_SIZE)
 	/* Legacy. -------------------------------------------------- */
-	return (size_t)sysconf( _SC_PHYS_PAGES ) *
-		(size_t)sysconf( _SC_PAGE_SIZE );
+	return (u_int64_t)sysconf( _SC_PHYS_PAGES ) *
+		(u_int64_t)sysconf( _SC_PAGE_SIZE );
 
 #elif defined(CTL_HW) && (defined(HW_PHYSMEM) || defined(HW_REALMEM))
 	/* DragonFly BSD, FreeBSD, NetBSD, OpenBSD, and OSX. -------- */
@@ -202,14 +202,14 @@ size_t rlib::getMemorySize( )
 #elif defined(HW_PYSMEM)
 	mib[1] = HW_PHYSMEM;		/* Others. ------------------ */
 #endif
-	unsigned int size = 0;		/* 32-bit */
+	uint32_t size = 0;          /* 32-bit */
 	size_t len = sizeof( size );
 	if ( sysctl( mib, 2, &size, &len, NULL, 0 ) == 0 )
-		return (size_t)size;
-	return 0L;			/* Failed? */
+		return (u_int64_t)size;
+	return (u_int64_t)0L;			/* Failed? */
 #endif /* sysctl and sysconf variants */
 
 #else
-	return 0L;			/* Unknown OS. */
+	return (u_int64_t)0L;			/* Unknown OS. */
 #endif
 }
